Fix: you don't need to add a 'display: hidden' property to body tag
The body tag contains all the HTML elements of a page. So you would not want to hide it as hiding the body tag might hide everything on your page. And there is no such css display property such as 'display: hidden'. I think you were looking for 'visibility: hidden'.